DG Updated Code-20240518 (1)/DG Updated Code-20240518/app/requirements.txt
azure-iot-device~=2.0.0
six